+++ tiled/client/__init__.py
@@ -1,9 +1,9 @@
-from ..utils import tree
 from .constructors import from_context, from_profile, from_uri
 from .container import ASCENDING, DESCENDING
 from .context import Context
 from .logger import hide_logs, record_history, show_logs
 from .metadata_update import DELETE_KEY
+from .utils import tree
 
 __all__ = [
     "ASCENDING",